Split up package install scripts into a builder abstraction
