+++ src/util/execute-lifecycle-script.js
@@ -6,7 +6,6 @@ import {MessageError, SpawnError} from '../errors.js';
 import * as constants from '../constants.js';
 import * as child from './child.js';
 import {registries} from '../resolvers/index.js';
-import {fixCmdWinSlashes} from './fix-cmd-win-slashes.js';
 
 const path = require('path');
 
@@ -108,9 +107,6 @@ export async function executeLifecycleScript(
   cmd: string,
   spinner?: ReporterSpinner,
 ): LifecycleReturn {
-  // if we don't have a spinner then pipe everything to the terminal
-  const stdio = spinner ? undefined : 'inherit';
-
   const env = await makeEnv(stage, cwd, config);
 
   // split up the path
@@ -129,40 +125,9 @@ export async function executeLifecycleScript(
   // join path back together
   env[constants.ENV_PATH_KEY] = pathParts.join(path.delimiter);
 
-  // get shell
-  const conf = {windowsVerbatimArguments: false};
-  let sh = 'sh';
-  let shFlag = '-c';
-  if (process.platform === 'win32') {
-    // cmd or command.com
-    sh = process.env.comspec || 'cmd';
-
-    // d - Ignore registry AutoRun commands
-    // s - Strip " quote characters from command.
-    // c - Run Command and then terminate
-    shFlag = '/d /s /c';
-
-    // handle windows run scripts starting with a relative path
-    cmd = fixCmdWinSlashes(cmd);
-
-    // handle quotes properly in windows environments - https://github.com/nodejs/node/issues/5060
-    conf.windowsVerbatimArguments = true;
-  }
-
-  const stdout = await child.spawn(sh, [shFlag, cmd], {cwd, env, stdio, ...conf}, (data) => {
-    if (spinner) {
-      const line = data.toString() // turn buffer into string
-        .trim() // trim whitespace
-        .split('\n') // split into lines
-        .pop() // use only the last line
-        .replace(/\t/g, ' '); // change tabs to spaces as they can interfere with the console
-
-      if (line) {
-        spinner.tick(line);
-      }
-    }
-  });
-
+  const stdout = await child.spawnWithTicker(cmd, cwd, env, spinner && ((data) => {
+    spinner && spinner.tick(data);
+  }));
   return {cwd, command: cmd, stdout};
 }
